add missing alt-attribute
Web Content Accessibility Guidelines (WCAG) 2.0, Level A: 1.1.1 Non-text
Content

All images need an `alt` attribute. If you do not supply an `alt`
attribute, that'll mean that users who ca not see the image will not
understand what the image conveys. Make sure that the `alt` attribute
value is a useful, concise, and clear description of the image.